+++ packages/integrations/src/index.ts
@@ -6,5 +6,6 @@ export { Ember } from './ember';
 export { ExtraErrorData } from './extraerrordata';
 export { ReportingObserver } from './reportingobserver';
 export { RewriteFrames } from './rewriteframes';
+export { Tracing } from './tracing';
 export { Transaction } from './transaction';
 export { Vue } from './vue';