+++ packages/opentracing/src/index.ts
@@ -1,3 +1,0 @@
-export { Span } from './span';
-export { Tracer } from './tracer';
-export { OpenTracingIntegration as Integration } from './opentracingintegration';