+++ packages/types/src/event.ts
@@ -3,7 +3,6 @@ import { Exception } from './exception';
 import { Request } from './request';
 import { SdkInfo } from './sdkinfo';
 import { Severity } from './severity';
-import { Span } from './span';
 import { Stacktrace } from './stacktrace';
 import { User } from './user';
 
@@ -33,7 +32,6 @@ export interface Event {
   tags?: { [key: string]: string };
   extra?: { [key: string]: any };
   user?: User;
-  spans?: Span[];
   type?: EventType;
 }