+++ packages/types/src/span.ts
@@ -1,5 +1,7 @@
-/** JSDoc */
+/** Span holding trace_id, span_id */
 export interface Span {
-  /** JSDoc */
+  /** Return a traceparent compatible header string */
+  toTraceparent(): string;
+  /** Convert the object to JSON */
   toJSON(): object;
 }